requirements.txt
requests>=2.31.0
argon2-cffi>=23.1.0
pywin32>=306
pystray>=0.19.5
Pillow>=10.0.0
psutil>=5.9.0
